+++ web/src/stores/general.ts
@@ -4,3 +4,5 @@ export const currentWorkspaceIdAtom = atomWithStorage<string | null>(
   'currentWorkspaceId',
   null,
 );
+
+export const isConnectedAtom = atomWithStorage<boolean>('isConnected', false);
